pyproject.toml
[project]
name = "x402-solana"
version = "0.1.0"
description = "x402 payment protocol implementation for Solana"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "DefiCodeSmith" }
]
requires-python = ">=3.10"
keywords = ["x402", "solana", "payments", "web3", "micropayments", "spl"]
dependencies = [
    "solana>=0.35.0",
    "solders>=0.18.5",
    "spl-token>=0.0.3",
    "pydantic>=2.10.0",
    "httpx>=0.27.0",
    "base58>=2.1.1",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
